Maybe this is a bad idea
Updated the DependencyManager to use reflection to parse requested
classes for dependencies in the phpdoc comments.  If the constructor for
the class has the @dmManaged tag then it will look for either a
@dmProvider or @dmRequires tag.  If a @dmProvider is defined, it will
attempt to find and instantiate the provider and call the "load" method
on it (assuming it's a standard Provider).

If there isn't a @dmProvider listed, it will look at all the @dmRequires
tags.  Each tag will define a class and the parameter it is assigned to.

Example:
/**
* @dmManaged
* @dmRequires Namespace\Path\To\Class $param
* @dmRequires Namespace\Path\To\Class\Also $alsoParam
*/
public function __constructor($param, $alsoParam)

The ordering of the tags isn't important, just that there is a tag
corresponding to each required parameter of the class.