DM tweaks
Changed tag syntax slightly (dependencies use @param instead of $param).
Facilitating the use of a provider as a dependency by just calling get
again instead of returning the object and by ensuring a new instance of
whatever existing class you request is created even if it's not
dmManaged.

This allows the system to create a new provider, then use the existing
provider logic to resolve that in place of the dependency.