Restructuring
Made a lot of changes regarding response generation and shifting of
responsibilities.  Avoided having to manually make html/json response
decisions.

You can specify the content-type in your route.  If it is json, all you
need to set in your controller is the model.

If the route is not json (well, only json and html are supported atm),
then if you don't have a view set in your controller it will just return
the model.  Otherwise, it will run through the view rendering steps.

The response will then json_encode it's body if it sees it's supposed to
be a json response.

Added a PDO initialization layer that pulls settings from the
configuration file.

Added route parameter matching for url parameters.

Added more functionality to the models.  Namely made them traversable
through IteratroAggregate, and JsonSerializable.  Added a collection
model which simply acts as a collection of other models or data types.